+++ test/unit/mongo_db_admin/code_coverage.sh
@@ -14,7 +14,6 @@ coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/defrag.py
 coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/get_log.py
 coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/help_message.py
 coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/main.py
-coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/process_mail.py
 coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/rotate.py
 coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/run_program.py
 coverage run -a --source=mongo_db_admin test/unit/mongo_db_admin/status.py